- Add hour display conditions for destination template. - Move destination body template image between location info and body.
